cifsd: move major tcp functions and structs to a dedicated header
glob.h must die.

The patch moves some of major TCP related structs and functions
to a dedicated header file - transport.h. This is the first in a
series of TCP layering effort patches.
