+++ fh.c
@@ -27,6 +27,7 @@
 #include "smb1pdu.h"
 #include "oplock.h"
 #include "buffer_pool.h"
+#include "transport.h"
 
 /**
  * alloc_fid_mem() - alloc memory for fid management
