+++ oplock.c
@@ -28,6 +28,7 @@
 #include "oplock.h"
 
 #include "buffer_pool.h"
+#include "transport.h"
 
 bool oplocks_enable = true;
 #ifdef CONFIG_CIFS_SMB2_SERVER
