+++ smb2ops.c
@@ -24,6 +24,8 @@
 #include "export.h"
 #include "smb2pdu.h"
 
+#include "transport.h"
+
 struct smb_version_values smb20_server_values = {
 	.version_string = SMB20_VERSION_STRING,
 	.protocol_id = SMB20_PROT_ID,
